refactor(trigger): extract mapSubscribersToJobs utility function and clean up TriggerBase implementation
