+++ libs/application-generic/src/usecases/trigger-base/trigger-base.usecase.ts
@@ -23,6 +23,7 @@ import { IProcessSubscriberBulkJobDto } from '../../dtos';
 import { SubscriberProcessQueueService } from '../../services/queues/subscriber-process-queue.service';
 import { buildUsageKey } from '../../services/cache/key-builders';
 import { CacheService, FeatureFlagsService } from '../../services';
+import { mapSubscribersToJobs } from '../../utils';
 
 export type BaseTriggerCommand = {
   environmentId: string;
@@ -88,50 +89,8 @@ export abstract class TriggerBase {
       return;
     }
 
-    const jobs = this.mapSubscribersToJobs(subscriberSource, subscribers, command);
+    const jobs = mapSubscribersToJobs(subscriberSource, subscribers, command);
 
     return await this.subscriberProcessQueueAddBulk(jobs);
-  }
-
-  protected mapSubscribersToJobs(
-    subscriberSource: SubscriberSourceEnum,
-    subscribers: { subscriberId: string; topics?: Pick<TopicEntity, '_id' | 'key'>[] }[] | ISubscribersDefine[],
-    command: BaseTriggerCommand
-  ): IProcessSubscriberBulkJobDto[] {
-    return subscribers.map((subscriber) => {
-      const job: IProcessSubscriberBulkJobDto = {
-        name: command.transactionId + subscriber.subscriberId,
-        data: {
-          environmentId: command.environmentId,
-          organizationId: command.organizationId,
-          userId: command.userId,
-          transactionId: command.transactionId,
-          identifier: command.identifier,
-          payload: command.payload,
-          overrides: command.overrides,
-          subscriber,
-          topics: subscriber.topics,
-          templateId: command.template._id,
-          _subscriberSource: subscriberSource,
-          requestCategory: command.requestCategory,
-          controls: command.controls,
-          bridge: {
-            url: command.bridgeUrl,
-            workflow: command.bridgeWorkflow,
-          },
-          environmentName: command.environmentName,
-        },
-        groupId: command.organizationId,
-      };
-
-      if (command.actor) {
-        job.data.actor = command.actor;
-      }
-      if (command.tenant) {
-        job.data.tenant = command.tenant;
-      }
-
-      return job;
-    });
   }
 }
